Castor3D: Moved OverlayRenderer files to Render folder, and split.
