+++ include/Core/Castor3D/Render/Overlays/OverlayRenderer.hpp
@@ -5,26 +5,20 @@ See LICENSE file in root folder
 #define ___C3D_OverlayRenderer_H___
 
 #include "Castor3D/Overlay/TextOverlay.hpp"
-#include "Castor3D/Render/RenderModule.hpp"
+#include "Castor3D/Render/Overlays/OverlayVertexBufferIndex.hpp"
+#include "Castor3D/Render/Overlays/OverlayVertexBufferPool.hpp"
 
-#include "Castor3D/Limits.hpp"
-#include "Castor3D/Buffer/GeometryBuffers.hpp"
-#include "Castor3D/Buffer/UniformBuffer.hpp"
 #include "Castor3D/Render/Passes/CommandsSemaphore.hpp"
 #include "Castor3D/Shader/Ubos/CameraUbo.hpp"
-#include "Castor3D/Shader/Ubos/OverlayUbo.hpp"
 
 #include <CastorUtils/Graphics/Size.hpp>
 
 #include <RenderGraph/FramePassTimer.hpp>
 
 #include <ashespp/Command/CommandBuffer.hpp>
-#include <ashespp/Descriptor/DescriptorSetLayout.hpp>
-#include <ashespp/Descriptor/DescriptorSetPool.hpp>
 #include <ashespp/Image/ImageView.hpp>
 #include <ashespp/Pipeline/GraphicsPipeline.hpp>
 #include <ashespp/Pipeline/PipelineLayout.hpp>
-#include <ashespp/Pipeline/PipelineVertexInputStateCreateInfo.hpp>
 #include <ashespp/Sync/Semaphore.hpp>
 
 namespace castor3d
@@ -33,49 +27,7 @@ namespace castor3d
 		: public castor::OwnedBy< RenderSystem >
 	{
 	public:
-		template< typename VertexT, uint32_t CountT >
-		struct VertexBufferIndexT;
-
-		template< typename VertexT, uint32_t CountT >
-		struct VertexBufferPoolT;
-		template< typename VertexT, uint32_t CountT >
-		using VertexBufferPoolPtrT = std::unique_ptr< VertexBufferPoolT< VertexT, CountT > >;
-
-		class Preparer
-			: public OverlayVisitor
-		{
-		public:
-			C3D_API Preparer( Preparer const & ) = delete;
-			C3D_API Preparer & operator=( Preparer const & ) = delete;
-			C3D_API explicit Preparer( OverlayRenderer & renderer
-				, RenderDevice const & device
-				, VkRenderPass renderPass
-				, VkFramebuffer framebuffer );
-			C3D_API Preparer( Preparer && rhs );
-			C3D_API Preparer & operator=( Preparer && rhs );
-			C3D_API ~Preparer()override;
-
-			void visit( PanelOverlay const & overlay )override;
-			void visit( BorderPanelOverlay const & overlay )override;
-			void visit( TextOverlay const & overlay )override;
-
-		private:
-			template< typename QuadT, typename OverlayT, typename VertexT, uint32_t CountT >
-			void doPrepareOverlay( RenderDevice const & device
-				, OverlayT const & overlay
-				, Pass const & pass
-				, std::map< size_t, VertexBufferIndexT< VertexT, CountT > > & overlays
-				, std::vector< VertexBufferPoolPtrT< VertexT, CountT > > & vertexBuffers
-				, std::vector < VertexT > const & vertices
-				, FontTextureSPtr fontTexture );
-
-		private:
-			OverlayRenderer & m_renderer;
-			RenderDevice const & m_device;
-			VkRenderPass m_renderPass;
-		};
-
-		friend class Preparer;
+		friend class OverlayPreparer;
 
 	public:
 		/**
@@ -122,7 +74,7 @@ namespace castor3d
 		 *\param[in]	renderPass	La passe de rendu.
 		 *\param[in]	framebuffer	Le framebuffer recevant le résultat.
 		 */
-		C3D_API Preparer beginPrepare( RenderDevice const & device
+		C3D_API OverlayPreparer beginPrepare( RenderDevice const & device
 			, VkRenderPass renderPass
 			, VkFramebuffer framebuffer );
 		/**
@@ -160,96 +112,18 @@ namespace castor3d
 		}
 		/**@}*/
 
-	private:
-		struct Configuration
-		{
-			castor::Point4f positionRatio;
-			castor::Point4i renderSizeIndex;
-		};
-
-		struct Pipeline
-		{
-			ashes::PipelineLayoutPtr pipelineLayout;
-			ashes::GraphicsPipelinePtr pipeline;
-		};
-
 	public:
-		struct OverlayRenderNode
-		{
-			Pipeline & pipeline;
-			Pass const & pass;
-		};
-
-	public:
-		struct OverlayGeometryBuffers
-		{
-			GeometryBuffers noTexture;
-			GeometryBuffers textured;
-		};
-
 		struct FontTextureDescriptorConnection
 		{
 			ashes::DescriptorSetPtr descriptorSet;
 			FontTexture::OnChanged::connection connection{};
 		};
 
-		template< typename VertexT, uint32_t CountT >
-		struct VertexBufferIndexT;
-
-		template< typename VertexT, uint32_t CountT >
-		struct VertexBufferPoolT
-		{
-			using MyBufferIndex = VertexBufferIndexT< VertexT, CountT >;
-			using Quad = std::array< VertexT, CountT >;
-
-			VertexBufferPoolT( Engine & engine
-				, std::string const & debugName
-				, RenderDevice const & device
-				, CameraUbo const & cameraUbo
-				, ashes::DescriptorSetLayout const & descriptorLayout
-				, ashes::PipelineVertexInputStateCreateInfo const & noTexDecl
-				, ashes::PipelineVertexInputStateCreateInfo const & texDecl
-				, uint32_t count );
-			VertexBufferIndexT< VertexT, CountT > allocate( OverlayRenderNode & node );
-			void deallocate( VertexBufferIndexT< VertexT, CountT > const & index );
-			void upload( ashes::CommandBuffer const & cb );
-
-			Engine & engine;
-			RenderDevice const & device;
-			ashes::BufferPtr< Configuration > overlaysData;
-			castor::ArrayView< Configuration > overlaysBuffer;
-			ashes::PipelineVertexInputStateCreateInfo const & noTexDeclaration;
-			ashes::PipelineVertexInputStateCreateInfo const & texDeclaration;
-			VertexBufferPoolUPtr buffer;
-			std::vector< ObjectBufferOffset > allocated;
-			ashes::DescriptorSetPoolPtr descriptorPool;
-			ashes::DescriptorSetPtr descriptorSet;
-
-		private:
-			ashes::DescriptorSetPtr doCreateDescriptorSet( CameraUbo const & cameraUbo
-				, ashes::DescriptorSetLayout const & descriptorLayout
-				, std::string const & debugName )const;
-		};
-
-		template< typename VertexT, uint32_t CountT >
-		struct VertexBufferIndexT
-		{
-			operator bool()const
-			{
-				return index != InvalidIndex;
-			}
-
-			VertexBufferPoolT< VertexT, CountT > & pool;
-			OverlayRenderNode & node;
-			uint32_t index;
-			OverlayGeometryBuffers geometryBuffers{};
-		};
-
-		using PanelVertexBufferPool = VertexBufferPoolT< OverlayCategory::Vertex, 6u >;
+		using PanelVertexBufferPool = OverlayVertexBufferPoolT< OverlayCategory::Vertex, 6u >;
+		using BorderPanelVertexBufferPool = OverlayVertexBufferPoolT< OverlayCategory::Vertex, 8u * 6u >;
+		using TextVertexBufferPool = OverlayVertexBufferPoolT< TextOverlay::Vertex, MaxOverlayCharsPerBuffer >;
 		using PanelVertexBufferIndex = PanelVertexBufferPool::MyBufferIndex;
-		using BorderPanelVertexBufferPool = VertexBufferPoolT< OverlayCategory::Vertex, 8u * 6u >;
 		using BorderPanelVertexBufferIndex = BorderPanelVertexBufferPool::MyBufferIndex;
-		using TextVertexBufferPool = VertexBufferPoolT< TextOverlay::Vertex, MaxOverlayCharsPerBuffer >;
 		using TextVertexBufferIndex = TextVertexBufferPool::MyBufferIndex;
 
 	private:
@@ -264,16 +138,16 @@ namespace castor3d
 			, Pass const & pass
 			, TextureLayout const & texture
 			, Sampler const & sampler );
-		Pipeline doCreatePipeline( RenderDevice const & device
+		OverlayPipeline doCreatePipeline( RenderDevice const & device
 			, VkRenderPass renderPass
 			, Pass const & pass
 			, ashes::PipelineShaderStageCreateInfoArray program
 			, TextureCombine const & texturesFlags
 			, bool text );
-		Pipeline & doGetPipeline( RenderDevice const & device
+		OverlayPipeline & doGetPipeline( RenderDevice const & device
 			, VkRenderPass renderPass
 			, Pass const & pass
-			, std::map< uint32_t, Pipeline > & pipelines
+			, std::map< uint32_t, OverlayPipeline > & pipelines
 			, bool text );
 		ashes::PipelineShaderStageCreateInfoArray doCreateOverlayProgram( RenderDevice const & device
 			, TextureCombine const & texturesFlags
@@ -298,8 +172,8 @@ namespace castor3d
 		castor::Size m_size;
 		std::map< Pass const *, OverlayRenderNode > m_mapPanelNodes;
 		std::map< Pass const *, OverlayRenderNode > m_mapTextNodes;
-		std::map< uint32_t, Pipeline > m_panelPipelines;
-		std::map< uint32_t, Pipeline > m_textPipelines;
+		std::map< uint32_t, OverlayPipeline > m_panelPipelines;
+		std::map< uint32_t, OverlayPipeline > m_textPipelines;
 		castor::String m_previousCaption;
 		bool m_sizeChanged{ true };
 		CameraUbo m_cameraUbo;
