+++ include/Core/Castor3D/Render/RenderWindow.hpp
@@ -8,7 +8,6 @@ See LICENSE file in root folder
 #include "Castor3D/Miscellaneous/MiscellaneousModule.hpp"
 
 #include "Castor3D/Event/UserInput/UserInputListener.hpp"
-#include "Castor3D/Overlay/OverlayRenderer.hpp"
 #include "Castor3D/Render/Passes/CommandsSemaphore.hpp"
 #include "Castor3D/Render/Passes/RenderQuad.hpp"
 #include "Castor3D/Render/ToTexture/RenderToTextureModule.hpp"
