+++ source/Core/Castor3D/Overlay/PanelOverlay.cpp
@@ -1,7 +1,6 @@
 #include "Castor3D/Overlay/PanelOverlay.hpp"
 
 #include "Castor3D/Engine.hpp"
-#include "Castor3D/Overlay/OverlayRenderer.hpp"
 #include "Castor3D/Overlay/Overlay.hpp"
 
 CU_ImplementCUSmartPtr( castor3d, PanelOverlay )
