+++ source/Core/Castor3D/Overlay/TextOverlay.cpp
@@ -4,9 +4,9 @@
 #include "Castor3D/Cache/OverlayCache.hpp"
 #include "Castor3D/Event/Frame/GpuFunctorEvent.hpp"
 #include "Castor3D/Material/Material.hpp"
-#include "Castor3D/Overlay/Overlay.hpp"
-#include "Castor3D/Overlay/OverlayRenderer.hpp"
 #include "Castor3D/Material/Texture/TextureLayout.hpp"
+#include "Castor3D/Overlay/Overlay.hpp"
+#include "Castor3D/Render/Overlays/OverlayRenderer.hpp"
 
 #include <CastorUtils/Graphics/Font.hpp>
 #include <CastorUtils/Graphics/Image.hpp>
