+++ source/Core/Castor3D/Render/Overlays/OverlayRenderer.cpp
@@ -1,12 +1,10 @@
-#include "Castor3D/Overlay/OverlayRenderer.hpp"
+/*
+See LICENSE file in root folder
+*/
+#include "Castor3D/Render/Overlays/OverlayRenderer.hpp"
 
 #include "Castor3D/Engine.hpp"
-#include "Castor3D/Limits.hpp"
 #include "Castor3D/Buffer/GpuBuffer.hpp"
-#include "Castor3D/Buffer/ObjectBufferPool.hpp"
-#include "Castor3D/Buffer/UniformBufferPool.hpp"
-#include "Castor3D/Buffer/PoolUniformBuffer.hpp"
-#include "Castor3D/Cache/MaterialCache.hpp"
 #include "Castor3D/Material/Material.hpp"
 #include "Castor3D/Material/Pass/Pass.hpp"
 #include "Castor3D/Material/Pass/Component/PassComponentRegister.hpp"
@@ -20,22 +18,19 @@
 #include "Castor3D/Overlay/TextOverlay.hpp"
 #include "Castor3D/Render/RenderNodesPass.hpp"
 #include "Castor3D/Render/RenderSystem.hpp"
+#include "Castor3D/Render/Overlays/OverlayPreparer.hpp"
 #include "Castor3D/Scene/Camera.hpp"
 #include "Castor3D/Shader/Program.hpp"
-#include "Castor3D/Shader/ShaderBuffers/PassBuffer.hpp"
-#include "Castor3D/Shader/ShaderBuffers/TextureAnimationBuffer.hpp"
-#include "Castor3D/Shader/ShaderBuffers/TextureConfigurationBuffer.hpp"
 #include "Castor3D/Shader/Shaders/GlslBlendComponents.hpp"
 #include "Castor3D/Shader/Shaders/GlslMaterial.hpp"
 #include "Castor3D/Shader/Shaders/GlslPassShaders.hpp"
 #include "Castor3D/Shader/Shaders/GlslTextureAnimation.hpp"
 #include "Castor3D/Shader/Shaders/GlslTextureConfiguration.hpp"
 #include "Castor3D/Shader/Shaders/GlslUtils.hpp"
+#include "Castor3D/Shader/Ubos/OverlayUbo.hpp"
 
 #include <CastorUtils/Graphics/Rectangle.hpp>
 #include <CastorUtils/Miscellaneous/Hash.hpp>
-
-#include <RenderGraph/FramePassTimer.hpp>
 
 #include <ashespp/RenderPass/FrameBuffer.hpp>
 
@@ -51,87 +46,6 @@ namespace castor3d
 
 	namespace ovrlrend
 	{
-		enum class OverlayBindingId : uint32_t
-		{
-			eMaterials,
-			eTexConfigs,
-			eTexAnims,
-			eMatrix,
-			eOverlays,
-		};
-
-		template< typename VertexT, uint32_t CountT >
-		static uint32_t doFillBuffers( typename std::vector< VertexT >::const_iterator begin
-			, uint32_t count
-			, OverlayRenderer::VertexBufferIndexT< VertexT, CountT > & bufferIndex )
-		{
-			auto & bufferOffset = bufferIndex.geometryBuffers.textured.bufferOffset;
-			std::copy( begin
-				, std::next( begin, count )
-				, bufferOffset.template getData< VertexT >( SubmeshFlag::ePositions ).begin() );
-			bufferOffset.markDirty( SubmeshFlag::ePositions );
-			return count;
-		}
-
-		template< typename ConfigT >
-		static void doUpdateUbo( ConfigT & data
-			, OverlayCategory const & overlay
-			, Pass const & pass
-			, castor::Size const & size )
-		{
-			auto position = overlay.getAbsolutePosition();
-			auto ratio = overlay.getRenderRatio( size );
-			data.positionRatio = castor::Point4f
-			{
-				position[0],
-				position[1],
-				ratio[0],
-				ratio[1],
-			};
-			data.renderSizeIndex = castor::Point4i
-			{
-				size.getWidth(),
-				size.getHeight(),
-				pass.getId(),
-				0,
-			};
-		}
-
-		template< typename OverlayT, typename QuadT >
-		struct BorderSizeGetter
-		{
-			castor::Rectangle operator()( OverlayT const & overlay, castor::Size const & size )const
-			{
-				return castor::Rectangle{};
-			}
-		};
-
-		template<>
-		struct BorderSizeGetter< BorderPanelOverlay, OverlayRenderer::BorderPanelVertexBufferPool::Quad >
-		{
-			castor::Rectangle operator()( BorderPanelOverlay const & overlay, castor::Size const & size )const
-			{
-				castor::Rectangle result = overlay.getAbsoluteBorderSize( size );
-
-				switch ( overlay.getBorderPosition() )
-				{
-				case BorderPosition::eMiddle:
-					result.set( result.left() / 2
-						, result.top() / 2
-						, result.right() / 2
-						, result.bottom() / 2 );
-					break;
-				case BorderPosition::eExternal:
-					break;
-				default:
-					result = castor::Rectangle{};
-					break;
-				}
-
-				return result;
-			}
-		};
-
 		enum class OverlayTexture : uint32_t
 		{
 			eNone = 0x00,
@@ -321,355 +235,6 @@ namespace castor3d
 
 	//*********************************************************************************************
 
-	OverlayRenderer::Preparer::Preparer( OverlayRenderer & renderer
-		, RenderDevice const & device
-		, VkRenderPass renderPass
-		, VkFramebuffer framebuffer )
-		: m_renderer{ renderer }
-		, m_device{ device }
-		, m_renderPass{ renderPass }
-	{
-		m_renderer.doBeginPrepare( m_renderPass, framebuffer );
-	}
-
-	OverlayRenderer::Preparer::Preparer( Preparer && rhs )
-		: m_renderer{ rhs.m_renderer }
-		, m_device{ rhs.m_device }
-		, m_renderPass{ rhs.m_renderPass }
-	{
-		rhs.m_renderPass = VkRenderPass{};
-	}
-
-	OverlayRenderer::Preparer & OverlayRenderer::Preparer::operator=( Preparer && rhs )
-	{
-		m_renderPass = rhs.m_renderPass;
-
-		rhs.m_renderPass = VkRenderPass{};
-
-		return *this;
-	}
-
-	OverlayRenderer::Preparer::~Preparer()
-	{
-		if ( m_renderPass )
-		{
-			m_renderer.doEndPrepare();
-		}
-	}
-
-	void OverlayRenderer::Preparer::visit( PanelOverlay const & overlay )
-	{
-		if ( auto material = overlay.getMaterial() )
-		{
-			for ( auto & pass : *material )
-			{
-				if ( !pass->isImplicit() )
-				{
-					doPrepareOverlay< PanelVertexBufferPool::Quad >( m_device
-						, overlay
-						, *pass
-						, m_renderer.m_panelOverlays
-						, m_renderer.m_panelVertexBuffers
-						, overlay.getPanelVertex()
-						, nullptr );
-				}
-			}
-		}
-	}
-
-	void OverlayRenderer::Preparer::visit( BorderPanelOverlay const & overlay )
-	{
-		if ( auto material = overlay.getMaterial() )
-		{
-			for ( auto & pass : *material )
-			{
-				if ( !pass->isImplicit() )
-				{
-					doPrepareOverlay< PanelVertexBufferPool::Quad >( m_device
-						, overlay
-						, *pass
-						, m_renderer.m_panelOverlays
-						, m_renderer.m_panelVertexBuffers
-						, overlay.getPanelVertex()
-						, nullptr );
-				}
-			}
-		}
-
-		if ( auto material = overlay.getBorderMaterial() )
-		{
-			for ( auto & pass : *material )
-			{
-				if ( !pass->isImplicit() )
-				{
-					doPrepareOverlay< BorderPanelVertexBufferPool::Quad >( m_device
-						, overlay
-						, *pass
-						, m_renderer.m_borderPanelOverlays
-						, m_renderer.m_borderVertexBuffers
-						, overlay.getBorderVertex()
-						, nullptr );
-				}
-			}
-		}
-	}
-
-	void OverlayRenderer::Preparer::visit( TextOverlay const & overlay )
-	{
-		if ( auto material = overlay.getMaterial() )
-		{
-			for ( auto & pass : *material )
-			{
-				if ( !pass->isImplicit() )
-				{
-					doPrepareOverlay< TextVertexBufferPool::Quad >( m_device
-						, overlay
-						, *pass
-						, m_renderer.m_textOverlays
-						, m_renderer.m_textVertexBuffers
-						, overlay.getTextVertex()
-						, overlay.getFontTexture() );
-				}
-			}
-		}
-	}
-
-	//*********************************************************************************************
-
-	namespace ovrlrend
-	{
-		template< typename VertexT, uint32_t CountT >
-		OverlayRenderer::VertexBufferIndexT< VertexT, CountT > & getVertexBuffer( std::vector< std::unique_ptr< OverlayRenderer::VertexBufferPoolT< VertexT, CountT > > > & pools
-			, std::map< size_t, OverlayRenderer::VertexBufferIndexT< VertexT, CountT > > & overlays
-			, Overlay const & overlay
-			, Pass const & pass
-			, OverlayRenderer::OverlayRenderNode & node
-			, Engine & engine
-			, RenderDevice const & device
-			, CameraUbo const & cameraUbo
-			, ashes::DescriptorSetLayout const & descriptorLayout
-			, ashes::PipelineVertexInputStateCreateInfo const & noTexLayout
-			, ashes::PipelineVertexInputStateCreateInfo const & texLayout
-			, uint32_t maxCount )
-		{
-			auto hash = std::hash< Overlay const * >{}( &overlay );
-			hash = castor::hashCombinePtr( hash, pass );
-			auto it = overlays.find( hash );
-
-			if ( it == overlays.end() )
-			{
-				for ( auto & pool : pools )
-				{
-					if ( it == overlays.end() )
-					{
-						auto result = pool->allocate( node );
-
-						if ( bool( result ) )
-						{
-							it = overlays.emplace( hash, std::move( result ) ).first;
-						}
-					}
-				}
-
-				if ( it == overlays.end() )
-				{
-					pools.emplace_back( std::make_unique< OverlayRenderer::VertexBufferPoolT< VertexT, CountT > >( engine
-						, getName( overlay.getType() )
-						, device
-						, cameraUbo
-						, descriptorLayout
-						, noTexLayout
-						, texLayout
-						, maxCount ) );
-					auto result = pools.back()->allocate( node );
-					it = overlays.emplace( hash, std::move( result ) ).first;
-				}
-			}
-
-			return it->second;
-		}
-	}
-
-	//*********************************************************************************************
-
-	template< typename QuadT, typename OverlayT, typename VertexT, uint32_t CountT >
-	void OverlayRenderer::Preparer::doPrepareOverlay( RenderDevice const & device
-		, OverlayT const & overlay
-		, Pass const & pass
-		, std::map< size_t, OverlayRenderer::VertexBufferIndexT< VertexT, CountT > > & overlays
-		, std::vector< OverlayRenderer::VertexBufferPoolPtrT< VertexT, CountT > > & vertexBuffers
-		, std::vector < VertexT > const & vertices
-		, FontTextureSPtr fontTexture )
-	{
-		if ( !vertices.empty() )
-		{
-			VertexBufferIndexT< VertexT, CountT > & bufferIndex = ovrlrend::getVertexBuffer< VertexT, CountT >( vertexBuffers
-				, overlays
-				, overlay.getOverlay()
-				, pass
-				, ( fontTexture
-					? m_renderer.doGetTextNode( device, m_renderPass, pass, *fontTexture->getTexture(), *fontTexture->getSampler().lock() )
-					: m_renderer.doGetPanelNode( device, m_renderPass, pass ) )
-				, *pass.getOwner()->getEngine()
-				, device
-				, m_renderer.m_cameraUbo
-				, *m_renderer.m_baseDescriptorLayout
-				, ( fontTexture ? m_renderer.m_noTexTextDeclaration : m_renderer.m_noTexDeclaration )
-				, ( fontTexture ? m_renderer.m_texTextDeclaration : m_renderer.m_texDeclaration )
-				, MaxOverlayPanelsPerBuffer );
-			ovrlrend::doUpdateUbo( bufferIndex.pool.overlaysBuffer[bufferIndex.index]
-				, overlay
-				, pass
-				, m_renderer.m_size );
-			ovrlrend::doFillBuffers( vertices.begin()
-				, uint32_t( vertices.size() )
-				, bufferIndex );
-			ashes::DescriptorSetCRefArray descriptorSets;
-			descriptorSets.push_back( *bufferIndex.pool.descriptorSet );
-			descriptorSets.push_back( *device.renderSystem.getEngine()->getTextureUnitCache().getDescriptorSet() );
-
-			if ( fontTexture )
-			{
-				descriptorSets.push_back( m_renderer.doCreateTextDescriptorSet( *fontTexture ) );
-			}
-
-			auto borderSize = ovrlrend::BorderSizeGetter< OverlayT, QuadT >{}( overlay, m_renderer.m_size );
-			auto borderOffset = castor::Size{ uint32_t( borderSize.left() ), uint32_t( borderSize.top() ) };
-			auto borderExtent = borderOffset + castor::Size{ uint32_t( borderSize.right() ), uint32_t( borderSize.bottom() ) };
-			auto position = overlay.getAbsolutePosition( m_renderer.m_size ) - borderOffset;
-			position->x = std::max( 0, position->x );
-			position->y = std::max( 0, position->y );
-			auto size = overlay.getAbsoluteSize( m_renderer.m_size ) + borderExtent;
-			size->x = std::max( 1u, size->x );
-			size->y = std::max( 1u, size->y );
-			ashes::CommandBuffer & commandBuffer = *m_renderer.m_commands.commandBuffer;
-			commandBuffer.bindPipeline( *bufferIndex.node.pipeline.pipeline );
-			commandBuffer.setViewport( makeViewport( m_renderer.m_size ) );
-			commandBuffer.setScissor( makeScissor( position, size ) );
-			commandBuffer.bindDescriptorSets( descriptorSets, *bufferIndex.node.pipeline.pipelineLayout );
-			DrawConstants constants{ bufferIndex.index };
-			commandBuffer.pushConstants( *bufferIndex.node.pipeline.pipelineLayout
-				, VK_SHADER_STAGE_VERTEX_BIT
-				, 0u
-				, sizeof( constants )
-				, &constants );
-			commandBuffer.bindVertexBuffer( 0u
-				, bufferIndex.geometryBuffers.noTexture.bufferOffset.getBuffer( SubmeshFlag::ePositions )
-				, bufferIndex.geometryBuffers.noTexture.bufferOffset.getOffset( SubmeshFlag::ePositions ) );
-			commandBuffer.draw( uint32_t( vertices.size() )
-				, 1u
-				, 0u
-				, 0u );
-		}
-	}
-
-	//*********************************************************************************************
-
-	template< typename VertexT, uint32_t CountT >
-	OverlayRenderer::VertexBufferPoolT< VertexT, CountT >::VertexBufferPoolT( Engine & engine
-		, std::string const & debugName
-		, RenderDevice const & device
-		, CameraUbo const & cameraUbo
-		, ashes::DescriptorSetLayout const & descriptorLayout
-		, ashes::PipelineVertexInputStateCreateInfo const & noTexDecl
-		, ashes::PipelineVertexInputStateCreateInfo const & texDecl
-		, uint32_t count )
-		: engine{ engine }
-		, device{ device }
-		, overlaysData{ makeBuffer< Configuration >( device
-			, MaxPipelines
-			, VK_BUFFER_USAGE_STORAGE_BUFFER_BIT
-			, VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT
-			, debugName ) }
-		, overlaysBuffer{ castor::makeArrayView( overlaysData->lock( 0u, ashes::WholeSize, 0u )
-			, overlaysData->getCount() ) }
-		, noTexDeclaration{ noTexDecl }
-		, texDeclaration{ texDecl }
-		, buffer{ castor::makeUnique< VertexBufferPool >( device, debugName ) }
-		, descriptorPool{ descriptorLayout.createPool( 1u ) }
-		, descriptorSet{ doCreateDescriptorSet( cameraUbo, descriptorLayout, debugName ) }
-	{
-	}
-
-	template< typename VertexT, uint32_t CountT >
-	OverlayRenderer::VertexBufferIndexT< VertexT, CountT > OverlayRenderer::VertexBufferPoolT< VertexT, CountT >::allocate( OverlayRenderNode & node )
-	{
-		auto it = std::find_if( allocated.begin()
-			, allocated.end()
-			, []( ObjectBufferOffset const & lookup )
-			{
-				return lookup.getBufferChunk( SubmeshFlag::ePositions ).buffer == nullptr;
-			} );
-
-		if ( it == allocated.end() )
-		{
-			allocated.emplace_back( buffer->getBuffer< Quad >( 1u ) );
-			it = std::next( allocated.begin(), ptrdiff_t( allocated.size() - 1u ) );
-		}
-
-		OverlayRenderer::VertexBufferIndexT< VertexT, CountT > result{ *this
-			, node
-			, uint32_t( std::distance( allocated.begin(), it ) ) };
-		ObjectBufferOffset & offsets = *it;
-
-		result.geometryBuffers.noTexture.bufferOffset = offsets;
-		result.geometryBuffers.noTexture.layouts.emplace_back( noTexDeclaration );
-
-		result.geometryBuffers.textured.bufferOffset = offsets;
-		result.geometryBuffers.textured.layouts.emplace_back( texDeclaration );
-
-		return result;
-	}
-
-	template< typename VertexT, uint32_t CountT >
-	void OverlayRenderer::VertexBufferPoolT< VertexT, CountT >::deallocate( OverlayRenderer::VertexBufferIndexT< VertexT, CountT > const & index )
-	{
-		CU_Require( &index.pool == this );
-		auto it = std::find_if( allocated.begin()
-			, allocated.end()
-			, [&index]( auto const & lookup )
-			{
-				return lookup.first == index.index;
-			} );
-		CU_Require( it != allocated.end() );
-		buffer->putBuffer( *it );
-		*it = {};
-		index.geometryBuffers.noTexture.bufferOffset = {};
-		index.geometryBuffers.noTexture.layouts.clear();
-		index.geometryBuffers.textured.bufferOffset = {};
-		index.geometryBuffers.textured.layouts.clear();
-	}
-
-	template< typename VertexT, uint32_t CountT >
-	void OverlayRenderer::VertexBufferPoolT< VertexT, CountT >::upload( ashes::CommandBuffer const & cb )
-	{
-		buffer->upload( cb );
-	}
-
-	template< typename VertexT, uint32_t CountT >
-	ashes::DescriptorSetPtr OverlayRenderer::VertexBufferPoolT< VertexT, CountT >::doCreateDescriptorSet( CameraUbo const & cameraUbo
-		, ashes::DescriptorSetLayout const & descriptorLayout
-		, std::string const & debugName )const
-	{
-		auto result = descriptorPool->createDescriptorSet( debugName );
-		engine.getMaterialCache().getPassBuffer().createBinding( *result
-			, descriptorLayout.getBinding( uint32_t( ovrlrend::OverlayBindingId::eMaterials ) ) );
-		engine.getMaterialCache().getTexConfigBuffer().createBinding( *result
-			, descriptorLayout.getBinding( uint32_t( ovrlrend::OverlayBindingId::eTexConfigs ) ) );
-		engine.getMaterialCache().getTexAnimBuffer().createBinding( *result
-			, descriptorLayout.getBinding( uint32_t( ovrlrend::OverlayBindingId::eTexAnims ) ) );
-		cameraUbo.createSizedBinding( *result
-			, descriptorLayout.getBinding( uint32_t( ovrlrend::OverlayBindingId::eMatrix ) ) );
-		result->createBinding( descriptorLayout.getBinding( uint32_t( ovrlrend::OverlayBindingId::eOverlays ) )
-			, *overlaysData
-			, 0u
-			, uint32_t( overlaysData->getCount() ) );
-		result->update();
-		return result;
-	}
-
-	//*********************************************************************************************
-
 	OverlayRenderer::OverlayRenderer( RenderDevice const & device
 		, Texture const & target
 		, crg::FramePassTimer & timer
@@ -708,13 +273,13 @@ namespace castor3d
 		std::string name = "Overlays";
 		auto & materials = getRenderSystem()->getEngine()->getMaterialCache();
 		ashes::VkDescriptorSetLayoutBindingArray baseBindings;
-		baseBindings.emplace_back( materials.getPassBuffer().createLayoutBinding( uint32_t( ovrlrend::OverlayBindingId::eMaterials ) ) );
-		baseBindings.emplace_back( materials.getTexConfigBuffer().createLayoutBinding( uint32_t( ovrlrend::OverlayBindingId::eTexConfigs ) ) );
-		baseBindings.emplace_back( materials.getTexAnimBuffer().createLayoutBinding( uint32_t( ovrlrend::OverlayBindingId::eTexAnims ) ) );
-		baseBindings.emplace_back( makeDescriptorSetLayoutBinding( uint32_t( ovrlrend::OverlayBindingId::eMatrix )
+		baseBindings.emplace_back( materials.getPassBuffer().createLayoutBinding( uint32_t( OverlayBindingId::eMaterials ) ) );
+		baseBindings.emplace_back( materials.getTexConfigBuffer().createLayoutBinding( uint32_t( OverlayBindingId::eTexConfigs ) ) );
+		baseBindings.emplace_back( materials.getTexAnimBuffer().createLayoutBinding( uint32_t( OverlayBindingId::eTexAnims ) ) );
+		baseBindings.emplace_back( makeDescriptorSetLayoutBinding( uint32_t( OverlayBindingId::eCamera )
 			, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER
 			, VK_SHADER_STAGE_VERTEX_BIT ) );
-		baseBindings.emplace_back( makeDescriptorSetLayoutBinding( uint32_t( ovrlrend::OverlayBindingId::eOverlays )
+		baseBindings.emplace_back( makeDescriptorSetLayoutBinding( uint32_t( OverlayBindingId::eOverlays )
 			, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER
 			, VK_SHADER_STAGE_VERTEX_BIT ) );
 		m_baseDescriptorLayout = device->createDescriptorSetLayout( name + "Base"
@@ -798,11 +363,11 @@ namespace castor3d
 		}
 	}
 
-	OverlayRenderer::Preparer OverlayRenderer::beginPrepare( RenderDevice const & device
+	OverlayPreparer OverlayRenderer::beginPrepare( RenderDevice const & device
 		, VkRenderPass renderPass
 		, VkFramebuffer framebuffer )
 	{
-		return Preparer{ *this, device, renderPass, framebuffer };
+		return OverlayPreparer{ *this, device, renderPass, framebuffer };
 	}
 
 	void OverlayRenderer::upload( ashes::CommandBuffer const & cb )
@@ -844,7 +409,7 @@ namespace castor3d
 		m_timerBlock = {};
 	}
 
-	OverlayRenderer::OverlayRenderNode & OverlayRenderer::doGetPanelNode( RenderDevice const & device
+	OverlayRenderNode & OverlayRenderer::doGetPanelNode( RenderDevice const & device
 		, VkRenderPass renderPass
 		, Pass const & pass )
 	{
@@ -859,7 +424,7 @@ namespace castor3d
 		return it->second;
 	}
 
-	OverlayRenderer::OverlayRenderNode & OverlayRenderer::doGetTextNode( RenderDevice const & device
+	OverlayRenderNode & OverlayRenderer::doGetTextNode( RenderDevice const & device
 		, VkRenderPass renderPass
 		, Pass const & pass
 		, TextureLayout const & texture
@@ -898,7 +463,7 @@ namespace castor3d
 		return *descriptorConnection.descriptorSet;
 	}
 
-	OverlayRenderer::Pipeline OverlayRenderer::doCreatePipeline( RenderDevice const & device
+	OverlayPipeline OverlayRenderer::doCreatePipeline( RenderDevice const & device
 		, VkRenderPass renderPass
 		, Pass const & pass
 		, ashes::PipelineShaderStageCreateInfoArray program
@@ -953,14 +518,14 @@ namespace castor3d
 				, ashes::PipelineDynamicStateCreateInfo{ 0u, { VK_DYNAMIC_STATE_VIEWPORT, VK_DYNAMIC_STATE_SCISSOR } }
 				, *pipelineLayout
 				, renderPass } );
-		return Pipeline{ std::move( pipelineLayout )
+		return OverlayPipeline{ std::move( pipelineLayout )
 			, std::move( pipeline ) };
 	}
 
-	OverlayRenderer::Pipeline & OverlayRenderer::doGetPipeline( RenderDevice const & device
+	OverlayPipeline & OverlayRenderer::doGetPipeline( RenderDevice const & device
 		, VkRenderPass renderPass
 		, Pass const & pass
-		, std::map< uint32_t, Pipeline > & pipelines
+		, std::map< uint32_t, OverlayPipeline > & pipelines
 		, bool text )
 	{
 		// Remove unwanted flags
@@ -999,10 +564,10 @@ namespace castor3d
 			VertexWriter writer;
 
 			C3D_Camera( writer
-				, ovrlrend::OverlayBindingId::eMatrix
+				, OverlayBindingId::eCamera
 				, 0u );
 			C3D_Overlays( writer
-				, ovrlrend::OverlayBindingId::eOverlays
+				, OverlayBindingId::eOverlays
 				, 0u );
 
 			sdw::PushConstantBuffer pcb{ writer, "C3D_DrawData", "c3d_drawData" };
@@ -1043,14 +608,14 @@ namespace castor3d
 				, utils };
 			shader::Materials materials{ writer
 				, passShaders
-				, uint32_t( ovrlrend::OverlayBindingId::eMaterials )
+				, uint32_t( OverlayBindingId::eMaterials )
 				, 0u };
 			shader::TextureConfigurations textureConfigs{ writer
-				, uint32_t( ovrlrend::OverlayBindingId::eTexConfigs )
+				, uint32_t( OverlayBindingId::eTexConfigs )
 				, 0u
 				, hasTexture };
 			shader::TextureAnimations textureAnims{ writer
-				, uint32_t( ovrlrend::OverlayBindingId::eTexAnims )
+				, uint32_t( OverlayBindingId::eTexAnims )
 				, 0u
 				, hasTexture };
