+++ source/Core/Castor3D/Render/RenderTarget.cpp
@@ -14,16 +14,15 @@
 #include "Castor3D/Miscellaneous/ProgressBar.hpp"
 #include "Castor3D/Overlay/Overlay.hpp"
 #include "Castor3D/Overlay/OverlayCategory.hpp"
-#include "Castor3D/Overlay/OverlayRenderer.hpp"
-#include "Castor3D/Render/RenderModule.hpp"
 #include "Castor3D/Render/RenderSystem.hpp"
-#include "Castor3D/Render/Culling/FrustumCuller.hpp"
-#include "Castor3D/Render/Node/SceneRenderNodes.hpp"
-#include "Castor3D/Render/Passes/OverlayPass.hpp"
-#include "Castor3D/Render/PostEffect/PostEffect.hpp"
 #include "Castor3D/Render/RenderTechnique.hpp"
 #include "Castor3D/Render/RenderTechniqueVisitor.hpp"
+#include "Castor3D/Render/Culling/FrustumCuller.hpp"
 #include "Castor3D/Render/EnvironmentMap/EnvironmentMap.hpp"
+#include "Castor3D/Render/Node/SceneRenderNodes.hpp"
+#include "Castor3D/Render/Overlays/OverlayRenderer.hpp"
+#include "Castor3D/Render/Passes/OverlayPass.hpp"
+#include "Castor3D/Render/PostEffect/PostEffect.hpp"
 #include "Castor3D/Render/ToneMapping/ToneMapping.hpp"
 #include "Castor3D/Render/ToTexture/Texture3DTo2D.hpp"
 #include "Castor3D/Scene/Camera.hpp"
